requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# RL
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0
gymnasium>=0.29.0

# NLP
transformers>=4.35.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Language detection
langdetect>=1.0.9
lingua-language-detector>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Discord bot
discord.py>=2.3.0

# Monitoring
wandb>=0.15.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Demo UI (HF Spaces)
gradio>=4.0.0
